metadati/requirements.txt
Flask
PyJWT
pymongo